+++ superset-frontend/packages/superset-ui-chart-controls/src/utils/index.ts
@@ -25,3 +25,4 @@ export { default as columnChoices } from './columnChoices';
 export * from './defineSavedMetrics';
 export * from './getStandardizedControls';
 export * from './getTemporalColumns';
+export * from './onChartClick';
