+++ superset-frontend/packages/superset-ui-core/src/utils/index.ts
@@ -31,4 +31,3 @@ export { getSelectedText } from './getSelectedText';
 export * from './featureFlags';
 export * from './random';
 export * from './typedMemo';
-export { onChartClickRedirectionHandler } from './onClickRedirectHandler';
