+++ superset-frontend/plugins/plugin-chart-echarts/src/Timeseries/EchartsTimeseries.tsx
@@ -21,7 +21,6 @@ import {
   DTTM_ALIAS,
   BinaryQueryObjectFilterClause,
   AxisType,
-  onChartClickRedirectionHandler,
 } from '@superset-ui/core';
 import { ViewRootGroup } from 'echarts/types/src/util/types';
 import GlobalModel from 'echarts/types/src/model/Global';
@@ -31,6 +30,7 @@ import Echart from '../components/Echart';
 import { TimeseriesChartTransformedProps } from './types';
 import { currentSeries } from '../utils/series';
 import { ExtraControls } from '../components/ExtraControls';
+import { onChartClickRedirectionHandler } from '@superset-ui/chart-controls';
 
 const TIMER_DURATION = 300;
